@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "7.2.1+insiders-2.12.0",
+  "version": "7.2.2+insiders-2.12.0",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -27,10 +27,13 @@
     "build": "rimraf material && ts-node -T tools/build --optimize",
     "build:dirty": "ts-node -T tools/build --dirty",
     "clean": "rimraf material",
-    "lint": "npm run lint:scss && npm run lint:ts",
+    "lint": "run-p lint:*",
     "lint:scss": "stylelint \"src/assets/**/*.scss\"",
     "lint:ts": "eslint --cache \"src/**/*.ts\"",
-    "start": "ts-node -T tools/build --verbose --dirty --watch"
+    "start": "ts-node -T tools/build --verbose --dirty --watch",
+    "upgrade": "run-s upgrade:*",
+    "upgrade:bump": "ncu --upgrade --filterVersion \"/^\\^/\"",
+    "upgrade:install": "npm install"
   },
   "dependencies": {
     "clipboard": "^2.0.8",
@@ -39,41 +42,43 @@
     "fuzzaldrin-plus": "^0.6.0",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.8.0",
-    "rxjs": "^7.2.0"
+    "rxjs": "^7.3.0"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^5.15.3",
     "@mdi/svg": "^5.9.55",
-    "@primer/octicons": "^14.2.2",
+    "@primer/octicons": "^15.0.0",
     "@types/clipboard": "^2.0.7",
     "@types/escape-html": "1.0.1",
     "@types/fuzzaldrin-plus": "^0.6.2",
     "@types/html-minifier": "^4.0.1",
     "@types/lunr": "^2.3.4",
-    "@types/node": "^16.4.3",
+    "@types/node": "^16.4.10",
     "@types/resize-observer-browser": "^0.1.6",
     "@types/sass": "^1.16.1",
-    "@typescript-eslint/eslint-plugin": "^4.28.4",
-    "@typescript-eslint/parser": "^4.28.4",
+    "@typescript-eslint/eslint-plugin": "^4.28.5",
+    "@typescript-eslint/parser": "^4.28.5",
     "autoprefixer": "^10.3.1",
     "chokidar": "^3.5.2",
     "cssnano": "^5.0.7",
-    "esbuild": "^0.12.16",
-    "eslint": "^7.31.0",
+    "esbuild": "^0.12.17",
+    "eslint": "^7.32.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsdoc": "^36.0.2",
+    "eslint-plugin-jsdoc": "^36.0.6",
     "eslint-plugin-no-null": "^1.0.2",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
     "html-minifier": "^4.0.0",
     "material-design-color": "^2.3.2",
     "material-shadows": "^3.0.1",
+    "npm-check-updates": "^11.8.3",
+    "npm-run-all": "^4.1.5",
     "postcss": "^8.3.6",
     "postcss-inline-svg": "^5.0.0",
     "preact": "^10.5.14",
     "rimraf": "^3.0.2",
-    "sass": "^1.36.0",
+    "sass": "^1.37.0",
     "stylelint": "^13.13.1",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-recommended": "^5.0.0",
